Lots of changes. Latest include updating things to work with new Trixie base. 	modified:   TODO 	new file:   apps-wip/shockos-imager-1.0-all/DEBIAN/control 	new file:   apps-wip/shockos-imager-1.0-all/README 	new file:   apps-wip/shockos-imager-1.0-all/TODO 	new file:   apps-wip/shockos-imager-1.0-all/usr/bin/shockos-imager 	new file:   apps-wip/shockos-imager-1.0-all/usr/share/glib-2.0/schemas/net.shockos.Imager.gschema.xml 	new file:   apps-wip/shockos-imager-1.0-all/usr/share/shockos/imager/main.ui 	new file:   apps-wip/shockos-imager-1.0-all/usr/share/shockos/imager/shockos-imager.cmb 	modified:   apps-wip/shockos-optimizations-1.0-all/usr/bin/shockos-optimizations 	modified:   apps-wip/shockos-optimizations-1.0-all/usr/share/shockos/optimizations/main.ui 	modified:   apps-wip/shockos-optimizations-1.0-all/usr/share/shockos/optimizations/shockos-optimizations.cmb 	new file:   apps-wip/shockos-welcome-1.0-all/main.ui 	new file:   apps-wip/shockos-welcome-1.0-all/shockos-welcome.cmb 	modified:   apps/MATE/shockos-mate-backgrounds-1.0-all/etc/systemd/user/shockos-mate-slideshow-background-daemon.service 	modified:   apps/MATE/shockos-mate-backgrounds-1.0-all/usr/bin/shockos-mate-backgrounds 	modified:   apps/shockos-update-manager-1.0-all/DEBIAN/postinst 	deleted:    apps/shockos-update-manager-1.0-all/etc/systemd/user/shockos-update-manager.service 	new file:   apps/shockos-update-manager-1.0-all/etc/xdg/autostart/net.shockos.UpdateManager.desktop 	modified:   apps/shockos-update-manager-1.0-all/usr/bin/shockos-update-manager 	new file:   apps/shockos-update-manager-1.0-all/usr/libexec/shockos/update-manager/check-for-updates-wrapper.sh 	modified:   apps/shockos-update-manager-1.0-all/usr/libexec/shockos/update-manager/check-for-updates.py 	modified:   build.sh 	new file:   python3-libs-wip/__pycache__/shocktools_download.cpython-312.pyc 	new file:   python3-libs-wip/shocktools_download.py 	modified:   shockos-core-10-all/DEBIAN/control 	modified:   shockos-core-10-all/DEBIAN/postinst 	modified:   shockos-core-10-all/usr/lib/shockos/dist-info 	modified:   shockos-core-10-all/usr/libexec/shockos/initsetup.sh 	new file:   shockos-core-10-all/usr/libexec/shockos/postinstall.sh 	modified:   shockos-core-10-all/usr/libexec/shockos/pyapt.py 	modified:   shockos-deskenv-mate-10-all/DEBIAN/control

+++ build.sh
@@ -16,7 +16,10 @@ then
       esac
     fi
     echo "SHOCKOS_BUILD_DATE=\"$(date "+%A, %B %e$suffix, %Y")\"" | tee ./shockos-core-10-all/usr/lib/shockos/dist-info
-    rm -r ./BUILD-OUT
+    if [[ -e ./BUILD_OUT ]]
+    then
+        rm -r ./BUILD-OUT
+    fi
     mkdir ./BUILD-OUT
     cp ./auto-install.sh ./BUILD-OUT/
     dpkg-deb --build ./shockos-core-10-all ./BUILD-OUT
